Changed the function return type from size_t to int

+++ Header Files/memset.h
@@ -6,6 +6,6 @@
 #include <stdio.h>
 
 /*Function Declaration Section*/
-size_t my_memset(char* str, int val, size_t n);
+int my_memset(char* str, int val, size_t n);
 
 #endif
